refactor: remove domain realm dependencies
