+++ .agents/memory-bank/activeContext.md
@@ -6,6 +6,7 @@ The current branch is `codex/realm-convergence`. Keep this file focused on the a
 ## Current Branch Projects
 These records stay active until the branch is ready to merge. Before merge, move completed project records into the matching history file and leave only the next focus, constraints, and history index here.
 
+- **Domain Realm Boundary Shrink (2026-07-03, complete):** Removed direct Realm usage from `Models/Domain/`. Deleted the now-empty `BookPreference.swift` after moving the pure `BookAnnotation.PrefId(...)` helper into `CalibreCoreModels.swift`; the per-server reader sidecar `Realm.Configuration`, migration block, schema object list, and Application Support file path now live in `DefaultServerScopedRealmConfigurationProvider`. `SortCriteria` is a pure domain enum again, with `PersistableEnum` conformance moved to `Models/Realm/CalibreSearchCache.swift`. AppContainer and reader preference repository fallbacks now resolve through the default provider instead of Domain. Validation: Domain Realm scans passed with no matches; focused `DefaultServerScopedRealmConfigurationProviderTests`, `ReaderPreferenceRepositoryTests`, `RealmReadingPositionRepositoryTests`, `UnifiedSearchServiceTests`, `UnifiedSearchIntegrationTests`, and `UnifiedCategoryServiceTests` passed (51 tests); after deleting `BookPreference.swift`, focused `ReaderPreferenceRepositoryTests` passed (14 tests); standard iOS Simulator build passed with existing warnings; `git diff --check` passed.
 - **LibrarySearchService Realm Boundary Shrink (2026-07-03, complete):** Removed direct `RealmSwift`, `Realm`, and `CalibreBookRealm` usage from `Models/Services/LibrarySearchService.swift`. Reused the existing `SearchCacheRepository` / `RealmSearchCacheStore` boundary instead of adding a new repository layer: the protocol now exposes value-type book lookup, local library search, and metadata-entry writes; `RealmSearchCacheStore` owns the Realm predicates, primary-key lookup, and metadata cache writes. Extracted Calibre metadata payload parsing into shared `CalibreBookMetadataValue` / `CalibreBook.applyMetadataValue(...)` in `Models/Calibre/CalibrePayloadModels.swift`; `CalibreBookRealm.applyMetadataEntry(...)` now delegates through the same value while keeping Realm-specific list/data writes in `Models/Realm/CalibreRealmMappers.swift`, and `CalibreServerService.handleLibraryBookOne(...)` no longer duplicates metadata field mapping. Added focused `UnifiedSearchIntegrationTests` coverage for local search filtering/sorting/pagination and metadata write/readback, plus `CalibreServerServiceTests` coverage that metadata refresh preserves existing format cache fields. Validation: `LibrarySearchService` Realm scans passed with no matches; duplicate metadata parser scan passed with no matches in network/Realm mapper files; focused `UnifiedSearchServiceTests`, `UnifiedSearchIntegrationTests`, `UnifiedCategoryServiceTests`, and `CalibreServerServiceTests.testGetMetadataAsyncSuccess` passed (25 tests); standard iOS Simulator build passed with existing warnings; `git diff --check` passed.
 - **CalibreActivityLogger Realm Boundary Shrink (2026-07-03, complete):** Removed direct `RealmSwift`, `Realm.Configuration`, and `CalibreActivityLogEntry` usage from `Network/CalibreActivityLogger.swift`. Added pure activity-log write values in `Models/Domain/ActivityLogWriteEvent.swift`; moved Realm object population/update matching into `Models/Realm/CalibreActivityLogRealmMappers.swift`; extended `ActivityLogRepositoryProtocol` / `RealmActivityLogRepository` as the read/write/delete/cleanup boundary for activity logs. `CalibreActivityLogger` now only batches request snapshots and delegates persistence to the repository. `AppContainer` and `DatabaseBootstrapper` construct the logger from `activityLogRepository`, and the now-unused `DatabaseService.loggerConfiguration()` helper was removed. While validating, fixed `RealmActivityLogRepository` to refresh before fetches and key its thread-local Realm cache by configuration to avoid stale in-memory Realm reuse across async tests. Also corrected `UnifiedSearchIntegrationTests` to build its test `CalibreServerService` with `container.databaseService` rather than `DatabaseService.shared`. Validation: logger Realm scans passed with no matches; focused `CalibreActivityLoggerTests`, `ActivityLogRepositoryTests`, `ActivityListViewModelTests`, `DatabaseBootstrapperTests`, `UnifiedSearchServiceTests`, and `UnifiedSearchIntegrationTests` passed (24 tests); standard iOS Simulator build passed; `git diff --check` passed.
 - **Book Detail Readium Nil Book Regression Fix (2026-07-03, complete):** Fixed a state mismatch where opening a book from Book Detail populated `ReadingSessionManager.readerInfo` but left `readingBook` nil or stale, which could surface as "Nil Book" when opening Readium EPUB or related reader sheets. `BookDetailViewModel` now routes read and preview-dismiss entry points through a shared helper that sets the active reading book via `CalibreBookManager`, allowing `ReadingSessionManager` to derive matching `readerInfo`. Added regression coverage to `BookDetailViewModelTests.testReadBookWhenInShelf` asserting the active reading book matches the selected detail book. Validation: focused `BookDetailViewModelTests`, `YabrReadiumReaderViewControllerTests`, `ReadiumPreferenceValueTests`, and `ReaderPreferenceRepositoryTests` passed (45 tests); `git diff --check` passed; standard iOS Simulator build passed.
